Refactoring project file structure
