+++ frontend/src/Pages/Main/Main.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import Divider from '@mui/material/Divider'
-import Markdown from './Markdown'
+import Markdown from '../../Components/Markdowns/Markdown'
 function Main ({ title }) {
   const [posts, setPosts] = useState([])
   useEffect(() => {
