+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './Assets/Styles/index.css';
 import { BrowserRouter } from 'react-router-dom'
-import App from './App';
+import App from './Routes';
 // import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
